requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
hyperlink==19.0.0
idna==2.8
imageio==2.5.0
incremental==17.5.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
jieba==0.39
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
klein==17.10.0
lesscpy==0.13.0
lxml==4.3.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
mongo==0.2.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.4
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
ply==3.11
prometheus-client==0.7.0
prompt-toolkit==2.0.9
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyExecJS==1.5.1
Pygments==2.4.2
PyHamcrest==1.9.0
pymongo==3.8.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypiwin32==223
pyrsistent==0.15.2
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywin32==224
pywinpty==0.5.5
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.5.1
queuelib==1.5.0
requestium==0.1.9
requests==2.22.0
requests-file==1.4.3
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.3.0
Scrapy==1.6.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
tensorboard==1.12.0
tensorflow==1.12.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
terminado==0.8.2
tesseract==0.1.3
testpath==0.4.2
tldextract==2.2.1
torch==1.0.1
torchvision==0.2.2.post3
tornado==6.0.2
traitlets==4.3.2
Twisted==19.2.0
urllib3==1.25.2
w3lib==1.20.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
zope.interface==4.6.0
